api/requirements.txt
openpyxl==3.1.2
fpdf2==2.8.3
